Adicionando Controller Advices e Exceptions

+++ src/main/java/com/ifsuldeminas/pas/bcc/gamestorewbe/model/services/impl/ClienteServiceImpl.java
@@ -1,8 +1,12 @@
-package com.ifsuldeminas.pas.bcc.gamestorewbe.services.impl;
+package com.ifsuldeminas.pas.bcc.gamestorewbe.model.services.impl;
 
-import com.ifsuldeminas.pas.bcc.gamestorewbe.entities.Cliente.Cliente;
-import com.ifsuldeminas.pas.bcc.gamestorewbe.repositories.ClienteRepository;
-import com.ifsuldeminas.pas.bcc.gamestorewbe.services.ClienteService;
+
+import com.ifsuldeminas.pas.bcc.gamestorewbe.model.domain.cliente.Cliente;
+import com.ifsuldeminas.pas.bcc.gamestorewbe.model.exceptions.cliente.ClienteNotFoundException;
+import com.ifsuldeminas.pas.bcc.gamestorewbe.model.repositories.ClienteRepository;
+import com.ifsuldeminas.pas.bcc.gamestorewbe.model.services.ClienteService;
+import lombok.extern.java.Log;
+import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,6 +15,8 @@ import java.util.Optional;
 
 @Service
 public class ClienteServiceImpl implements ClienteService {
+
+    private static final Logger LOG = org.slf4j.LoggerFactory.getLogger(ClienteServiceImpl.class);
 
     @Autowired
     private ClienteRepository clienteRepository;
@@ -21,19 +27,22 @@ public class ClienteServiceImpl implements ClienteService {
     }
 
     @Override
-    public Cliente buscarClientePorId(Integer id) {
+    public Cliente buscarClientePorId(Integer id) throws ClienteNotFoundException {
         Optional<Cliente> cliente = clienteRepository.findById(id);
+        if (!cliente.isPresent()){
+            throw new ClienteNotFoundException(id);
+        }
         return cliente.orElse(null);
     }
 
     @Override
     public void addCliente(Cliente cliente) {
-        cliente.setIdCliente(null);
         clienteRepository.save(cliente);
+        LOG.info("Cliente adicionado com sucesso");
     }
 
     @Override
-    public void atualizaCliente(Cliente cliente) {
+    public void atualizaCliente(Cliente cliente) throws ClienteNotFoundException{
         Cliente atual = this.buscarClientePorId(cliente.getIdCliente());
         atual.setNome(cliente.getNome());
         atual.setCpf(cliente.getCpf());
@@ -41,10 +50,15 @@ public class ClienteServiceImpl implements ClienteService {
         atual.setDataNasc(cliente.getDataNasc());
         atual.setTelefone(cliente.getTelefone());
         clienteRepository.save(atual);
+        LOG.info("Cliente atualizado com sucesso");
     }
 
     @Override
-    public void deletaCliente(Integer id) {
+    public void deletaCliente(Integer id) throws ClienteNotFoundException{
+        if (!this.clienteRepository.existsById(id)){
+            throw new ClienteNotFoundException(id);
+        }
         clienteRepository.deleteById(id);
+        LOG.info("Cliente deletado com sucesso");
     }
 }
